+++ packages/product-feed-zbozi/CHANGELOG.md
@@ -5,12 +5,17 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
 - New dependencies for dev (@MattCzerner)
     - phpunit/phpunit >=5.0.0,<6.0
     - twig/twig 1.34.0
     - twig/extensions 1.3.0
 - New automatic test that is controlling right behaviour of plugin (@MattCzerner)
 - Added travis build icon into [README.md](README.md) (@MattCzerner)
+### Changed
+- Dependency [plugin-interface](shopsys/product-feed-interface) upgraded from ~0.2.0 to ~0.3.0 (@MattCzerner)
+### Removed
+- `ZboziFeedConfig::getFeedItemRepository()` (@MattCzerner)
 
 ## [0.3.0] - 2017-09-06
 ### Added